Remove redundant type cast

+++ app/src/main/java/com/example/yjh/areyouready/TouchImageView.java
@@ -1131,7 +1131,7 @@ public class TouchImageView extends android.support.v7.widget.AppCompatImageView
                 minY = maxY = startY;
             }
 
-            scroller.fling(startX, startY, (int) velocityX, (int) velocityY, minX,
+            scroller.fling(startX, startY, velocityX, velocityY, minX,
                     maxX, minY, maxY);
             currX = startX;
             currY = startY;
